requirements.txt
requests
pandas
scikit-learn
scipy
gradio
python-dotenv

